pyproject.toml
[project]
name = "rheology-ai-agent"
version = "0.1.0"
description = "Autonomous Rheological Control Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.3",
    "langgraph>=1.0.9",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
]

[tool.uv]
package = true